% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_years}
\alias{water_years}
\title{Water Year Index and Type 1901-2016}
\format{A data frame with 227 rows and 7 variables:
\describe{
\item{WY}{October-September}
\item{Oct_Mar}{}
\item{Apr_Jul}{}
\item{WYsum}{}
\item{Index}{}
\item{Yr_type}{}
\item{location}{Sacramento Valley or San Joaquin Valley}
}}
\source{

}
\usage{
water_years
}
\description{
blaw
}
\keyword{datasets}
